refactor: merge ContextDataRepository and CachedContextDataManager into one class
