+++ aggregator/src/main/java/io/spoud/kcc/aggregator/stream/MetricEnricher.java
@@ -9,6 +9,7 @@ import io.quarkus.logging.Log;
 import io.spoud.kcc.aggregator.CostControlConfigProperties;
 import io.spoud.kcc.aggregator.data.RawTelegrafData;
 import io.spoud.kcc.aggregator.olap.AggregatedMetricsRepository;
+import io.spoud.kcc.aggregator.repository.ContextDataRepository;
 import io.spoud.kcc.aggregator.repository.GaugeRepository;
 import io.spoud.kcc.aggregator.repository.MetricNameRepository;
 import io.spoud.kcc.aggregator.stream.serialization.SerdeFactory;
@@ -38,7 +39,7 @@ public class MetricEnricher {
     private static final ObjectMapper OBJECT_MAPPER =
             new ObjectMapper().registerModule(new JavaTimeModule());
     private final MetricNameRepository metricRepository;
-    private final CachedContextDataManager cachedContextDataManager;
+    private final ContextDataRepository contextDataRepository;
     private final CostControlConfigProperties configProperties;
     private final SerdeFactory serdes;
     private final GaugeRepository gaugeRepository;
@@ -132,7 +133,7 @@ public class MetricEnricher {
                 .setValue(telegrafData.getValue())
                 .setTags(rawTelegrafData.tags())
                 .setContext(Collections.emptyMap());
-        telegrafData.enrichWithContext(cachedContextDataManager).ifPresent(aggregatedDataInfo -> aggregatedData
+        telegrafData.enrichWithContext(contextDataRepository).ifPresent(aggregatedDataInfo -> aggregatedData
                 .setName(aggregatedDataInfo.name())
                 .setContext(aggregatedDataInfo.context())
                 .setEntityType(aggregatedDataInfo.type()));
@@ -247,7 +248,7 @@ public class MetricEnricher {
         if (principalName == null) {
             return metric;
         }
-        var newContext = cachedContextDataManager.getContextDataForName(EntityType.PRINCIPAL,
+        var newContext = contextDataRepository.getContextDataForName(EntityType.PRINCIPAL,
                 principalName, metric.getTimestamp());
         newContext.put("topic", metric.getName());
         return new AggregatedData(metric.getTimestamp(), EntityType.PRINCIPAL, principalName,
